work-bus/executor: don't commit to a work request until we're done
That is, extend the window in which we will actively keep a work request
message live across the entire job's execution, through completion
message receipt (or timeout).  By contrast, we used to commit as soon as
the board had booted but before the actual work had started.  While one
might have hoped, that's not sufficient, and we have observed a number
of crashes (especially, SATA bus explosions) between the `login:` prompt
and actually getting the GitHub action runner off the ground.

This way, if the board explodes before we are able to contact the
originating requestor (i.e., GitHub) for the job, we won't lose the work
request.  On the other hand, if we crash *while* running the job, it's
possible that the originating requestor will have already committed to
this board's runner being the one and only that could ever take this
job.  Oh well.

We rely on the delivery count limit on the service bus itself to limit
the damage that a job could do -- it will eventually fail to redeliver
and the job will be lost rather than cycling through the cluster
forever.  In the case where the requestor commits to a runner which then
crashes, this does mean we'll cycle a board up and then time out while
waiting for work, but that should constitute successful completion of
the job and we'll pull the work request off the bus.