sphinx: include local/**
It's probably a good idea for us to tidy up the local documentation such
that Sphinx can read it.  We'd like to make this conditional, but it's
not entirely obvious how to make it all hang together, since sphinx so
aggressively pulls on toctree directives (in particular, they cannot be
switched off with ifconfig or similar).