work-bus: upgrade to @azure/identity 3.1.1
After recent refactoring of the github-reflector to persist connections
to Azure resources, we were seeing access to these objects fail.  Others
had reported exactly the same problem in
https://github.com/Azure/azure-sdk-for-js/issues/22620
which turns out to be a special case of
https://github.com/Azure/azure-sdk-for-js/issues/22722
which is, in turn, fixed with
https://github.com/Azure/azure-sdk-for-js/pull/23232
which was part of the @azure/identity 3.0.0 release.

Jump up to 3.1.1 since it's the latest at the moment.