cheribsd-exports: github runner as non-root user
Create the user in /opt/bin/msr-morello-init (rather than requiring a
FreeBSD host box to change /etc), drop to that user in
/opt/bin/msr-github-act-runner.sh (to run
msr-github-act-runner-inner.sh).

We don't change the executor so that it remains compatible with our
older root images that don't play this game, but those are not long for
the world.