An example subset of local configuration
Some of this might be useful at your site, but much of it is just what happened
to be running on our cluster supervisory node (paths and such are a little too
hard-coded, for example), so expect that you'll need to make extensive
modifications.  Still, a worked example might be useful.